+++ internal/caller/servicemeta.go
@@ -10,29 +10,23 @@ import (
 	rpb "google.golang.org/grpc/reflection/grpc_reflection_v1alpha"
 )
 
-type ServiceMetaData interface {
-	GetServiceMetaDataList(context.Context) ([]*ServiceMeta, error)
-}
-
 type serviceMetaData struct {
-	connFact *rpc.GrpcConnFactory
-	target   string
-	deadline int
-}
+	connFact     *rpc.GrpcConnFactory
+	target       string
+	deadline     int
+	protoImports []string
 
-type ServiceMeta struct {
-	Name    string
-	Methods []*desc.MethodDescriptor
-	File    *desc.FileDescriptor
+	serviceMetaBase
 }
 
 // NewServiceMetaData returns new instance of ServiceMetaData
 // that reads service metadata by calling grpc Reflection service of the target
-func NewServiceMetaData(connFact *rpc.GrpcConnFactory, target string, deadline int) ServiceMetaData {
+func NewServiceMetaData(connFact *rpc.GrpcConnFactory, target string, deadline int, protoImports []string) ServiceMetaData {
 	return &serviceMetaData{
-		connFact: connFact,
-		target:   target,
-		deadline: deadline,
+		connFact:     connFact,
+		target:       target,
+		deadline:     deadline,
+		protoImports: protoImports,
 	}
 }
 
@@ -87,4 +81,8 @@ func (s *serviceMetaData) GetServiceMetaDataList(ctx context.Context) ([]*Servic
 
 	defer rc.Reset()
 	return res, nil
+}
+
+func (s *serviceMetaData) GetAdditionalFiles() ([]*desc.FileDescriptor, error) {
+	return s.serviceMetaBase.GetAdditionalFiles(s.protoImports)
 }
